% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abort_education.R
\name{abort_education}
\alias{abort_education}
\title{abort_education}
\usage{
abort_education(x)
}
\arguments{
\item{x}{str. Level of education from the education column in the dataset 'canada_ab'}
}
\value{
table
}
\description{
Computes the absolute frequency of the view on abortion given the education level from the dataset 'canada_ab'
}
\examples{
\dontrun{
abort_education(x='college')
}

}
