% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{agilent44k2ensg_chr}
\alias{agilent44k2ensg_chr}
\title{Biomart conversion}
\usage{
agilent44k2ensg_chr(probe_id, biomart = mart, combine = F, df2 = "",
  by.x = "efg_agilent_wholegenome_4x44k_v1",
  by.y = "efg_agilent_wholegenome_4x44k_v1", all = F)
}
\description{
Biomart conversion
}
\examples{
agilent44k2ensg_chr()
}
