% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_module.R
\name{analyze_ehr_data}
\alias{analyze_ehr_data}
\title{Analyze EHR Data}
\usage{
analyze_ehr_data(ehr_data)
}
\arguments{
\item{ehr_data}{A dataframe containing preprocessed EHR data.}
}
\description{
This function performs basic analysis on EHR data, such as summary statistics and distribution plots.
}
