% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{antidepressant}
\alias{antidepressant}
\title{antidepressant.
A data set containing  antidepressant trial data as described in paper by White,Royes,Best}
\format{
dataframe containing 688 rows and 14 columns
\describe{
\item{PATIENT.NUMBER}{}
\item{HAMA.TOTAL}{}
\item{PGI_IMPROVEMENT}{}
\item{VISIT...VISIT.3.DATE}{}
\item{VISIT.NUMBER}{}
\item{TREATMENT.NAME}{}
\item{PATIENT.SEX}{}
\item{POOLED.INVESTIGATOR}{}
\item{basval}{}
\item{HAMD17.TOTAL}{}
\item{change}{}
\item{miss_flag}{}
\item{methodcol}{}
\item{referencecol}{}
}
}
\usage{
antidepressant
}
\description{
antidepressant.
A data set containing  antidepressant trial data as described in paper by White,Royes,Best
}
\keyword{datasets}
