\name{assembleOutput}
\alias{assembleOutput}
\title{Internal function used for assembling the output produced by the EMR process...}
\usage{assembleOutput(myTempDirOut)}
\description{Internal function used for assembling the output produced by the EMR process}
\details{Internal function used for assembling the output produced by the EMR process}
\value{Returns a list of output that matches the input list in length}
\author{James "JD" Long}
\arguments{\item{myTempDirOut}{the temp directory where the output files from the EMR
process are kept. Must be local, not S3.}}
