% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_csr_props.R
\name{calculate_csr_props}
\alias{calculate_csr_props}
\title{Calculate CSR Strategy Proportions}
\usage{
calculate_csr_props(csr_data)
}
\arguments{
\item{csr_data}{DataFrame containing CSR values by site}
}
\value{
DataFrame with calculated CSR proportions
}
\description{
Calculate CSR Strategy Proportions
}
