% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_invasion_risk.R
\name{calculate_invasion_risk}
\alias{calculate_invasion_risk}
\title{Calculate Invasion Risk Score
#' @importFrom FD dbFD}
\usage{
calculate_invasion_risk(fd_indices, csr_props)
}
\arguments{
\item{fric}{Functional richness value}

\item{feve}{Functional evenness value}

\item{fdiv}{Functional divergence value}

\item{c_prop}{C-strategy proportion}

\item{s_prop}{S-strategy proportion}

\item{r_prop}{R-strategy proportion}
}
\value{
A numeric value between 0 and 1 representing invasion risk
}
\description{
Calculate Invasion Risk Score
#' @importFrom FD dbFD
}
\examples{
calculate_invasion_risk(0.5, 0.6, 0.4, 30, 40, 30)
}
