% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_analysis_functions.R
\name{cc_d.llk}
\alias{cc_d.llk}
\title{Function to calculate d(log likelihood)/db0
for logistic model given b1, b2, data}
\usage{
cc_d.llk(b0, b1, b2, aff, unaf)
}
\arguments{
\item{b0}{input}

\item{b1}{input}

\item{b2}{input}

\item{aff}{input}

\item{unaf}{input}
}
\description{
Function to calculate d(log likelihood)/db0
for logistic model given b1, b2, data
}
\examples{
cc_d.llk()
}
