% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.ntp.R
\name{cell_type_NTP}
\alias{cell_type_NTP}
\title{cell_type_NTP}
\usage{
cell_type_NTP(
  seurat,
  wd,
  markerList,
  assay = "RNA",
  slot = c("scale.data", "count.data", "data"),
  output.dir = "./",
  rda.dir = "./data",
  NTP.g.filter.method = c("sd", "mad", "none"),
  NTP.gene.filter.cutoff = 0.3,
  NTP.distance = c("cosine", "correlation"),
  NTP.norm.method = c("none", "row.std"),
  mc.cores = 1
)
}
\arguments{
\item{seurat}{Seurat object}

\item{wd}{working directory}

\item{markerList}{List of cell type marker}

\item{assay}{Assay to use}

\item{slot}{seurat object expression data, c("scale.data", "count.data", "data")()}

\item{output.dir}{output directory}

\item{rda.dir}{Path of the RData saving directory}

\item{NTP.g.filter.method}{Method of gene filtering in NTP c(sd (Default), mad, none)}

\item{NTP.gene.filter.cutoff}{Cut-off score of standard deviation in NTP}

\item{NTP.distance}{Method of calculating distance in NTP, either c("correlation" or "cosine").}

\item{NTP.norm.method}{Method of normalization in NTP, either c("none", "row.std")}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
Seurat object
}
\description{
A function to run NTP to infer cell type
}
\details{
cell type annotation using NTP
}
