\name{checkLastStepStatus}
\alias{checkLastStepStatus}
\title{AWS Support Function: Checks the status of a given job on EMR...}
\usage{checkLastStepStatus(jobFlowId)}
\description{AWS Support Function: Checks the status of a given job on EMR}
\details{Checks the status of a previously issued step.}
\value{Status of the last step}
\author{James "JD" Long}
\arguments{\item{jobFlowId}{the Job Flow Id of the job to check}}
