\name{checkStatus}
\alias{checkStatus}
\title{AWS Support Function: Checks the status of a given job on EMR...}
\usage{checkStatus(jobFlowId)}
\description{AWS Support Function: Checks the status of a given job on EMR}
\details{Checks the status of a previously issued job.}
\value{Job Status}
\author{James "JD" Long}
\arguments{\item{jobFlowId}{the Job Flow Id of the job to check}}
