% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.post.hoc.filas.R
\encoding{UTF-8}
\name{chisq.post.hoc.filas}
\alias{chisq.post.hoc.filas}
\title{Contrastes post hoc de una matriz de datos con chi-cuadrada}
\usage{
chisq.post.hoc.filas(tabla, p.adjust.method = NULL, alternative = NULL)
}
\arguments{
\item{tabla}{Una tabla o matriz, la misma que se usó con la función chisq.test}

\item{p.adjust.method}{Método usado para ajustar el valor de P. Se puede usar: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", y "none".}

\item{alternative}{Un argumento que indica la hipótesis alternativa que se probará.}
}
\value{
Contrastes 'post-hoc' de chi cuadrada para dos o mas filas usando la prueba Fisher.test.
}
\description{
Muestra las comparaciones post-hoc por pares de una prueba Chi cuadrada.
Realiza la prueba exacta de Fisher para comprobar independencia de filas en una tabla de contingencia, luego de obtener una prueba chi-cuadrada significativa.
}
\details{
Función tomada y modificada de la paqueteria rstatix. https://cran.r-project.org/web/packages/rstatix/index.html
}
\examples{
xtab <- as.table(rbind(c(180, 145), c(179, 106),c(510, 196), c(862, 23)))
dimnames(xtab) <- list(
Tratamiento = c("1st", "2nd", "3rd", "Control"),
Genero = c("hembra", "macho"))
chisq.post.hoc.filas(xtab)
#solo hipótesis alternativa
chisq.post.hoc.filas(xtab, alternative = "less")
chisq.post.hoc.filas(xtab, alternative = "greater")
chisq.post.hoc.filas(xtab, alternative = "two.sided")
#con los dos argumentos
chisq.post.hoc.filas(xtab, p.adjust.method = "BY", alternative = "less")
chisq.post.hoc.filas(xtab, p.adjust.method = "BY", alternative = "two.sided")
chisq.post.hoc.filas(xtab, p.adjust.method = "holm", alternative = "less")
}
