% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CIR_loop}
\alias{CIR_loop}
\title{CIR_loop}
\usage{
CIR_loop(c_mata_miss, mata_Means, MeansC)
}
\arguments{
\item{c_mata_miss}{vector of col locaton of missing values , eg 5 6}

\item{mata_Means}{vector of means after mcmc draws eg 17 1 16.8 15.5 14.6 13.2}

\item{MeansC}{vector of means after mcmc draws using variance from reference group}
}
\value{
mata_means
}
\description{
process CIR method
}
\details{
This is based on Suzie Cro's Stata program
}
