% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_articles.R
\name{clean_html_text}
\alias{clean_html_text}
\title{Clean HTML text}
\usage{
clean_html_text(value)
}
\arguments{
\item{value}{Value or list of values to clean.}
}
\value{
Character scalar with normalized text.
}
\description{
Converts possibly nested HTML-like values into normalized plain text.
}
