% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vparty.R
\name{convert_version}
\alias{convert_version}
\title{Convert regular V-Dem data release version to V-Party version}
\usage{
convert_version(datarelease, vparty_releases)
}
\arguments{
\item{datarelease}{Vector (character) with data releases information.}

\item{options}{Vector (character or numeric) of regular data releases in which V-Party data is released as well.}
}
\value{
Character vector of \code{datarelease} length with replacements.
}
\description{
Replace original data releases with V-Party releases
}
\examples{
\dontrun{
VPARTY_RELEASES <- unlist(strsplit("10.12.14", "[.]"))
datarelease <- c("10-14.",
"10-14. Available upon request, subject to review and approval",
"10-14.", "14.", "11.", "9-12", "9-10.")
convert_version(datarelease, VPARTY_RELEASES)
}

}
