% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CRB_wide}
\alias{CRB_wide}
\title{Wide Format Fish Returns at Bonneville Dam}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 87 rows and 10 columns.
}
\usage{
CRB_wide
}
\description{
A wide dataset containing annual counts of fish returns at Bonneville Dam in
the Columbia River Basin.
}
\keyword{datasets}
