% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.inferCNV.R
\name{cts.GO}
\alias{cts.GO}
\title{cts.GO}
\usage{
cts.GO(cell.type.set, rda.dir,cts.geneClustList)
}
\arguments{
\item{cell.type.set}{cell type set}

\item{rda.dir}{rData directory}

\item{cts.geneClustList}{cell type specific geneClustList}
}
\value{
gprofiler result List
}
\description{
A wrapper function to cts.GO
}
\details{
Interface to the g:Profiler tool for finding enrichments in gene lists.
}
