% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currcodes.R
\docType{data}
\name{currcodes}
\alias{currcodes}
\title{All currency codes}
\format{
A data frame with columns:
\describe{
 \item{Entity}{Country Names}
 \item{Currency}{Currency Names}
 \item{AlphabeticCode}{Currency Code}
 \item{NumericCode}{Numeric code of the currency}
 \item{MinorUnit}{decimals}
 \item{WithdrawalDate}{Withdrawal Date}
 }
}
\source{
\url{https://datahub.io/core/currency-codes}, downloaded 2020-12-16
}
\usage{
currcodes
}
\description{
Data frame of all currency names and codes.
}
\examples{
 currcodes

}
\keyword{datasets}
