% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tree.R
\encoding{UTF-8}
\name{data.tree}
\alias{data.tree}
\title{Graficar estructura anidada de los datos}
\usage{
data.tree(
  x,
  left = 0,
  top,
  right = 1,
  lastcenter = NA,
  showval = TRUE,
  showcount = TRUE,
  stacklabels = TRUE,
  firstcall = TRUE,
  col = NULL,
  border = NA,
  toplab = NULL,
  base.cex = 1,
  ...
)
}
\arguments{
\item{x}{Dataframe en el que cada columna sucesiva representa subcategorías de la columna anterior.}

\item{left}{El borde izquierdo de los rectángulos apilados en unidades especificadas por el usuario.}

\item{top}{La parte superior de los rectángulos apilados en unidades especificadas por el usuario.}

\item{right}{El borde derecho de la pila actual de rectángulos en unidades especificadas por el usuario.}

\item{lastcenter}{El centro del rectángulo anterior del que surge el siguiente desglose de categorías. No hay casi ninguna razón para cambiarlo.}

\item{showval}{Si se muestran los valores que representan las categorías.}

\item{showcount}{Si se desea mostrar el recuento de las categorías.}

\item{stacklabels}{Si se desea mostrar los nombres del dataframe debajo de los rectángulos apilados.}

\item{firstcall}{indicador para la función - no lo modifique.}

\item{col}{Colores de relleno opcionales para los rectángulos.}

\item{border}{Color para el borde alrededor de los rectángulos.}

\item{toplab}{Etiquetas opcionales para mostrar en la parte superior de cada pila de rectangulos.}

\item{base.cex}{Tamano de los caracteres base para las etiquetas.}

\item{...}{argumentos adicionales pasados a 'plot'.}
}
\value{
Grafico de tipo arbol que muestra la estructura de los datos.
}
\description{
Exploracion visual de la estructura de los datos. Permite analizar la estructura de anidamiento de los datos y el tamano de muestra por cada nivel. Util para establecer el anidamiento en el componente aleatorio de los modelos mixtos LMM, GLMM y GAM. Funcion tomada de la paqueteria "Plotrix".
Cita del paquete: J L (2006). “Plotrix: a package in the red light district of R.” R-News, 6(4), 8-12.
}
\examples{
data(sleep)
data.tree(sleep[c(2:3)])
}
