% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fit2.R
\name{data_fit2}
\alias{data_fit2}
\title{data_fit2}
\usage{
data_fit2(y_sim, x, D, psi, k, parallel = F, ncores = 1, type_1 = T)
}
\arguments{
\item{y_sim}{A matrix of dimension n * m corresponding to the normalized OTU table, with n denoting the number of subjects(patients), and m denoting the number of taxa.}

\item{x}{A matrix of dimension n * p corresponding to the meta data matrix, with n denotes the number of subjects(patients), and p denotes the number of covariates (age/BMI/...).}

\item{D}{A matrix of dimension m * m corresponding to the taxa distance matrix.}

\item{psi}{A scalar corresponding to the importance weight of the taxa distance matrix.}

\item{k}{A scalar corresponding to the number of nearest taxa in a phylogenetic tree we will use to impute a missing value. Theoretically, the larger k, the more accurate our imputation
will be (required that k <= m).}

\item{parallel}{A boolean indicating whether to use parallel or not. Default is FALSE.}

\item{ncores}{A scalar corresponding to the number of cores to use. It is only used when parallel = TRUE. Default is 1.}

\item{type_1}{A boolean indicates the type of imputation.}
}
\value{
A list containing two elements: 1. y_imp: the imputed submatrix 2. mse: the mse calculated based on the values that are not regarded as false values.
}
\description{
data_fit2 is a function that aims to perform one imputation in a group of biological samples, it's a subfunction of mbImpute and intend not to be
used by the user.
}
