% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_remove.R
\name{data_remove}
\alias{data_remove}
\title{Remover columnas}
\usage{
data_remove(data, select, ignore_case = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{data}{datos}

\item{select}{seleccionar}

\item{ignore_case}{Falso por default}

\item{verbose}{Falso por default}
}
\value{
Un dataframe sin las columnas
}
\description{
Esta funcion fue tomada de la paquetería DataWizard, revisar para detalles.
}
\keyword{internal}
