% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva.R
\name{decis}
\alias{decis}
\title{Decis da Distribuição}
\usage{
decis(x, decimais = 2)
}
\arguments{
\item{x}{Vetor numérico.}

\item{decimais}{Casas decimais (padrão 2).}
}
\description{
Decis da Distribuição
}
