% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coders.R
\name{delete_coder_question}
\alias{delete_coder_question}
\title{Delete a question from \code{coder_table}}
\usage{
delete_coder_question(question_id, db)
}
\arguments{
\item{question_id}{A V-Dem \code{question_id}}

\item{db}{A dataset construction database connection}
}
\value{
Returns NULL, invisibly.
}
\description{
Delete a question from \code{coder_table}
}
\details{
This function deletes a question from the \code{coder_table} given a valid question id.  It is used in the dataset construction module \code{gen_coder_tables.R}.
}
\examples{
\dontrun{
db <- vbase::pg_connect("v13")
delete_coder_question(2500, db) 
}

}
