\name{deleteS3Bucket}
\alias{deleteS3Bucket}
\title{AWS Support Function: Delete an S3 Bucket...}
\usage{deleteS3Bucket(bucketName)}
\description{AWS Support Function: Delete an S3 Bucket}
\details{Does nothing if the bucketName does not exist. If bucket contains Keys,
all keys are deleted.}
\author{James "JD" Long}
\arguments{\item{bucketName}{the bucket to be deleted}}
