\name{deleteS3Key}
\alias{deleteS3Key}
\title{AWS Support Function: Delete an S3 Key (a...}
\usage{deleteS3Key(bucketName, keyName)}
\description{AWS Support Function: Delete an S3 Key (a.k.a file)}
\details{Deteles a key in a given bucket on S3}
\author{James "JD" Long}
\arguments{\item{bucketName}{name of the bucket}
\item{keyName}{the key in the bucket}}
