% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva.R
\name{descrever}
\alias{descrever}
\title{Descrição Estatística Completa}
\usage{
descrever(x, por = NULL, numericas = TRUE, categoricas = TRUE, decimais = 2)
}
\arguments{
\item{x}{Um vetor numérico, vetor categórico ou data.frame.}

\item{numericas}{Lógico. Se TRUE e x for um data.frame, exibe resumo das variáveis numéricas.}

\item{categoricas}{Lógico. Se TRUE e x for um data.frame, exibe resumo das variáveis categóricas.}

\item{decimais}{Quantidade de casas decimais para exibição (padrão 2).}
}
\description{
Exibe resumos estatísticos detalhados para vetores numéricos, categóricos
ou para todo um conjunto de dados (data.frame).
}
