% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detachAllPackages.R
\encoding{UTF-8}
\name{detach_paquetes}
\alias{detach_paquetes}
\title{Desmontar paquetes cargados en el espacio de trabajo}
\usage{
detach_paquetes()
}
\value{
Quita todos lo paquetes cargados.
}
\description{
Desmontar todos los paquetes en R. Quita los paquetes cargados en el espacio de trabajo. Útil cuando
tenemos las mismas funciones con mismo nombre de diferentes paquetes.
}
\examples{
#detach_paquetes()
}
