% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_summarise.R
\name{dist_summary}
\alias{dist_summary}
\title{Summarise posterior distributions}
\usage{
dist_summary(m, expanded.names = NULL, drop.vignettes = T, utable = NULL, ...)
}
\arguments{
\item{m}{Matrix, either a coerced parameter from a \code{stanfit} object
or output from the \code{osp} or \code{ord} transformation
functions. Columns correspond to country-dates.}

\item{expanded.names}{Row names to expand the summarised time
series to using \code{\link{stretch}}.}

\item{drop.vignettes}{Boolean. Whether to drop vignette rows prior
to summarisation. Note, if set to \code{FALSE}, but
\code{expanded.names} is not \code{NULL}, the function will
return an error since vignette identifiers are not valid
country-dates.}

\item{...}{Additional arguments passed to the \code{\link{stretch}}
function.}
}
\value{
DataFrame containing the mean, median, sd,
codehigh/low68, codehigh/low95.
}
\description{
\code{dist_summary} summarises a country-date matrix where each column
is a posterior distribution.
}
\details{
\code{dist_summary} will preserve entire columns of
\code{NA} under the assumption that they're missing
country-dates; however, it will fail when only a single value
within a column is \code{NA} since we never want partial
missingness within our posterior distribution.
}
