% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\docType{data}
\name{edges}
\alias{edges}
\title{edges of phylogenetic tree of Karlsson et al.}
\format{
A data frame containing 145 rows and 12 columns, the first column indicates the condition we are interested in.
\describe{
\item{x}{1}
\item{y}{2}
}
}
\usage{
edges
}
\description{
The edges for 344 taxa
}
\keyword{datasets}
