\name{emptyS3Bucket}
\alias{emptyS3Bucket}
\title{AWS Support Function: Empty an S3 bucket...}
\usage{emptyS3Bucket(bucketName)}
\description{AWS Support Function: Empty an S3 bucket}
\details{Deletes all keys in the designated bucket}
\author{James "JD" Long}
\arguments{\item{bucketName}{Name of the bucket to be emptied}}
