% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{ensg2chr_start_end}
\alias{ensg2chr_start_end}
\title{Biomart conversion}
\usage{
ensg2chr_start_end(ensg, biomart = mart, combine = F, df2 = "",
  by.x = "ensembl_gene_id", by.y = "ensembl_gene_id", all = F,
  as.IGV = F)
}
\description{
Biomart conversion
}
\examples{
ensg2chr_start_end()
}
