% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{enst2ext_name_biotype}
\alias{enst2ext_name_biotype}
\title{Biomart conversion}
\usage{
enst2ext_name_biotype(ensg, biomart = mart, combine = F, df2 = "",
  by.x = "ensembl_transcript_id", by.y = "ensembl_transcript_id", all = F)
}
\description{
Biomart conversion
}
\examples{
enst2ext_name_biotype()
}
