% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{EvalPlot}
\alias{EvalPlot}
\title{Function automatically builds calibration plots for model evaluation}
\usage{
EvalPlot(data, PredColName = "PredictedValues",
  ActColName = "ActualValues", type = c("calibration", "boxplot"),
  bucket = 0.05, aggrfun = function(x) mean(x, na.rm = TRUE))
}
\arguments{
\item{data}{Data containing predicted values and actual values for comparison}

\item{PredColName}{String representation of column name with predicted values from model}

\item{ActColName}{String representation of column name with actual values from model}

\item{type}{Calibration or boxplot - calibration aggregated data based on summary statistic; boxplot shows variation}

\item{bucket}{Number of buckets to partition the space on (0,1) for evaluation}

\item{aggrfun}{The statistics function used in aggregation, listed as a function}
}
\value{
Calibration plot or boxplot
}
\description{
This function automatically builds calibration plots and calibration boxplots for model evaluation using regression, quantile regression, and binary and multinomial classification
}
\examples{
out1 <- EvalPlot(calibration,
                 PredColName = "preds",
                 ActColName  = "actuals",
                 type        = "calibration",
                 bucket      = 0.05,
                 aggrfun     = function(x) quantile(x, probs = 0.50, na.rm = TRUE))

out2 <- EvalPlot(calibration,
                 PredColName = "preds",
                 ActColName  = "actuals",
                 type        = "boxplot",
                 bucket      = 0.05)
}
\author{
Adrian Antico at RemixInstitute.com
}
