\name{fancyRpartPlot}
\alias{fancyRpartPlot}
\title{
  A wrapper for plotting rpart trees using prp
}
\description{
  
  Plots a fancy RPart decision tree using the pretty rpart plotter.
  
}
\usage{
fancyRpartPlot(model, main="", ...)

}
\arguments{
  
  \item{model}{an rpart object.}

  \item{main}{title for the plot.}

  \item{...}{additional arguments passed on to prp.}

}
\references{Package home page: \url{http://rattle.togaware.com}}
\author{\email{Graham.Williams@togaware.com}}
\keyword{hplot}
