% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_script.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Ploting map
This function uses names, generated by function make_filename to (in case of existense)
 import .csv files  by fars_read funnction.
With the specified state number all accidents are plotted on the map.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integer, that identifies state to plot observations on a map}

\item{year}{A parameter which identifies the value of year, used in built file name}
}
\value{
If specified state number is not presented in initial data, function stops with the corresponding message.
If there is such a state number in data, but there is no observations, nothing is plotted (with the printed message
about it) . Finally, if there are observations for specified state number, they are plotted on a map.
}
\description{
Ploting map
This function uses names, generated by function make_filename to (in case of existense)
 import .csv files  by fars_read funnction.
With the specified state number all accidents are plotted on the map.
}
\examples{
fars_map_state(1, 2011)
fars_map_state(2, 2013)is missing name/title. Skipping

}
