% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runmimix.R
\name{fillinterims}
\alias{fillinterims}
\title{fillinterims}
\usage{
fillinterims(impdata, interims, Mimp = M)
}
\arguments{
\item{impdata}{the data with missing values}

\item{interims}{the interim cases with estimated MAR values}

\item{Mimp}{the number of imputations specified , ie M total imputsations}
}
\value{
list of 1st data set with interims imputed plus M interim cases of each interim case to be matche in 2nd pass
}
\description{
fills missing interims distinguising from post-discontinuation
}
\details{
checks methodindiv not null
}
