% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findcurr.R
\name{findcurr}
\alias{findcurr}
\title{Find Currency Function}
\usage{
findcurr(text)
}
\arguments{
\item{text}{a character vector of text to search for in the country names}
}
\value{
a data frame of matched country names with their respective currency codes
}
\description{
a function to search for a string in the country names and return with their respective currency names and currency codes
}
\examples{
findcurr('united')
}
