% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findstock_av.R
\name{findstock_av}
\alias{findstock_av}
\title{Find Stock Symbol Function}
\usage{
findstock_av(text)
}
\arguments{
\item{text}{text/string to search in stock name}
}
\value{
data frame object of all matching stock names and their symbols
}
\description{
API wrapper to auto-complete stock names and search for its symbol
}
\examples{
findstock_av('mic')
}
