% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p_value.R
\docType{data}
\name{format_p_values}
\alias{format_p_values}
\alias{format_p_values.default}
\alias{format_p}
\alias{format_p_values.data.frame}
\alias{add_signif_stars}
\alias{get_signif_stars}
\alias{signif_syms_05s}
\alias{p05}
\alias{signif_syms_01s}
\alias{p01}
\alias{signif_syms_001s}
\alias{p001}
\alias{signif_syms_001}
\alias{signif_syms_01}
\alias{signif_syms_05}
\alias{signif_syms}
\alias{p05_01_001}
\alias{p05plus}
\alias{signif_stars_legend_2}
\alias{signif_stars_legend}
\alias{rm_zero}
\alias{format_as_p_columns}
\title{Formato de valores de p}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 2.

An object of class \code{numeric} of length 3.

An object of class \code{numeric} of length 4.

An object of class \code{numeric} of length 4.

An object of class \code{numeric} of length 4.
}
\usage{
format_p_values(
  p,
  digits_p = 3,
  cols = NULL,
  ...,
  alpha = 0.05,
  signif_stars = TRUE,
  rm_zero = FALSE,
  add_p = FALSE,
  rm_spaces = FALSE,
  ss = signif_syms
)

\method{format_p_values}{default}(
  p,
  digits_p = 3,
  cols = NULL,
  ...,
  signif_stars = TRUE,
  rm_zero = FALSE,
  add_p = FALSE,
  rm_spaces = FALSE,
  ss = signif_syms
)

format_p(
  p_i,
  digits_p = 3,
  signif_stars = TRUE,
  rm_zero = FALSE,
  add_p = FALSE,
  rm_spaces = FALSE,
  ss = signif_syms
)

\method{format_p_values}{data.frame}(
  p,
  digits_p = 3,
  cols = NULL,
  ...,
  ss = signif_syms,
  signif_stars = TRUE,
  rm_zero = FALSE,
  add_p = FALSE,
  rm_spaces = FALSE
)

add_signif_stars(p, ss = signif_syms)

get_signif_stars(p, ss = signif_syms)

signif_syms_05s

p05

signif_syms_01s

p01

signif_syms_001s

p001

signif_syms_001

signif_syms_01

signif_syms_05

signif_syms

p05_01_001

p05plus

signif_stars_legend_2(ss = signif_syms)

signif_stars_legend(
  ss = signif_syms,
  decreasing = FALSE,
  collapse = c("  \\n", ", ", "; ")
)

rm_zero(str, dec = ".")

format_as_p_columns(
  data,
  colnames = c("p.value", "p.adjust"),
  digits_p = 3,
  rm_zero = FALSE,
  signif_stars = FALSE,
  ...
)
}
\arguments{
\item{p}{Un (vector de) valores de p. Numerico.
O un \code{data.frame}.}

\item{digits_p}{(numerico) Número de dígitos significativos para redondear un valor p.
No menos de 2. \itemize{
\item if \code{digits_p = 2}: \enumerate{
\item los valores inferiores a 0.001 se imprimen como \code{"<0.001"};
\item valores entre 0.001 y 0.01 como \code{"<0.01"};
\item todos los demás valores se redondean a dos decimales.
}
\item if \code{digits_p = 3}, sólo formato \code{"<0.01"} se salta.
}}

\item{cols, alpha, ss, decreasing, collapse, dec}{\verb{# FIXME:} \emph{not documented}}

\item{...}{Argumentos para otros métodos.}

\item{signif_stars}{(logical) Marca si se deben añadir estrellas de significación a cada valor p.
No menos de 2.}

\item{rm_zero}{(logical) Marca si el cero inicial antes del punto debe ser removido.}

\item{add_p}{(logical) Marca si la letra "p" debe incluirse en la expresión.}

\item{rm_spaces}{(logical) Marca si se deben eliminar todos los espacios.}
}
\value{
Un vector de caracteres con valores p formateados.
}
\description{
Functions to fromat p values.
}
\details{
\itemize{
\item \code{format_p} - formatos único valor p.
\item \code{format_as_p_columns} - formatea las columnas numericas indicadas en un marco de datos como valores p (las columnas se convierten en cadenas).
\item \code{get_signif_stars} -  toma los valores numéricos de p aporta las estrellas de significación estadística apropiadas.
\item \code{add_signif_stars} - formatea los valores p numéricos añadiendo estrellas de significación (el resultado es un vector de caracteres).
\item \code{signif_stars_legend} - genera una leyenda para las estrellas de significación (el resultado es una cadena).
\item \code{rm_zero} - elimina el cero al principio de un número (devuelve una cadena con el mismo valor pero sin el cero inicial).
}
}
\note{
Función tomada y traducida de la paqueteria Biostat: https://github.com/GegznaV/biostat.
}
\examples{
# Prettify p-values

#format_p_values(0.0005)

#format_p_values(0.005)
#format_p_values(0.005, signif_stars = FALSE)
#format_p_values(0.005, rm_zero = TRUE)
#format_p_values(0.005, digits_p = 2)
#format_p_values(0.005, digits_p = 2, rm_zero = TRUE, signif_stars = FALSE)

#format_p_values("0.00022")
#format_p_values("0.052")

#format_p_values(c(0.005, 0.0005, 0.052147))


#get_signif_stars(0.005)
#add_signif_stars(0.005)

#get_signif_stars(0.0005)
#add_signif_stars(0.0005)

#get_signif_stars(0.052147)
#add_signif_stars(0.052147)

#signif_stars_legend()

#format_p(0)
#format_p(.02, digits_p = 2)
#format_p(.0002)
#format_p(.0002, signif_stars = FALSE)
#format_p(.0002, ss = c("*****" = 0.001))

# TODO:
# 1. Añadir parámetro para activar la corrección del valor p
#    de p = 1 a, por ejemplo, p > 0,999;
#
# 2. fusionar parámetros `ss` y `signif_stars`
# 3. test: format_p(NaN) -- [OK]
# 4. test: format_p(NA) -- esta función falla con NA como entrada.
rm_zero(0.020)
}
\keyword{datasets}
\keyword{internal}
