% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{genline2dosage}
\alias{genline2dosage}
\title{genline2dosage}
\usage{
genline2dosage(line, ind.offset = 0)
}
\arguments{
\item{line}{input}

\item{ind.offset}{input}
}
\description{
genline2dosage
}
\examples{
genline2dosage()

}
