\name{get_sss_codes}
\alias{get_sss_codes}
\title{Reads all "codes" inside the triple-s "record"...}
\usage{get_sss_codes(x)}
\description{Reads all "codes" inside the triple-s "record"}
\details{This function parses the record node and extracts all "codes" nodes into a data.frame}
\keyword{XML}
\arguments{\item{x}{XML node}}
