% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_tags.R
\name{get_survey}
\alias{get_survey}
\title{Get survey from question tag}
\usage{
get_survey(x)
}
\arguments{
\item{x}{CharacterVector of question tags.}
}
\description{
Given a vector of question tags, \code{get_survey} returns the
two letter survey abbreviation for A*, A, B, and C-variables
}
\examples{
get_survey(c("v2clacfree", "v2dlcountr"))

}
