% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getexchrate_av.R
\name{getexchrate_av}
\alias{getexchrate_av}
\title{Get Exchange Rate Function}
\usage{
getexchrate_av(to_curr, from_curr = "USD")
}
\arguments{
\item{to_curr}{(character) currency code of the buying currency}

\item{from_curr}{(character) currency code of the selling currency}
}
\value{
data.frame object of daily exchange rate  historic stock price
}
\description{
API wrapper to extract exchange rate data from www.alphavantage.co
}
\examples{
getexchrate_av('GBP','USD')
}
