% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getstockprice_av.R
\name{getstockprice_av}
\alias{getstockprice_av}
\title{Get Stock Price Function (AV)}
\usage{
getstockprice_av(symbol)
}
\arguments{
\item{symbol}{(character) currency code of the buying currency}
}
\value{
data.frame object of historic daily stock price
}
\description{
API wrapper to extract daily stock price data from www.alphavantage.co
}
\examples{
getstockprice_av('AAPL')
}
