% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{getUpstream}
\alias{getUpstream}
\title{get upstream sequence}
\usage{
getUpstream(input, bp_upstream = 2000, which.mart = mart,
  which.type = "external_gene_name", which.seq.type = "gene_flank",
  fasta.output = "")
}
\description{
get upstream sequence
}
\examples{
getUpstream()
}
