% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{goid2goname}
\alias{goid2goname}
\title{Biomart conversion}
\usage{
goid2goname(goid, biomart = mart, combine = F, df2 = "", by.x = "go_id",
  by.y = "go_id", all = F)
}
\description{
Biomart conversion
}
\examples{
goid2goname()
}
