% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{imprint}
\alias{imprint}
\title{Returns a logical matrix where the elements that are not NA in the
columns subsetted by the given logical vector are set as true.}
\usage{
imprint(m, v)
}
\arguments{
\item{m}{A numeric matrix}

\item{v}{A logical vector the same length as \code{ncol(m)}}
}
\description{
Returns a logical matrix where the elements that are not NA in the
columns subsetted by the given logical vector are set as true.
}
