% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_tags.R
\name{is.contemp}
\alias{is.contemp}
\alias{is.hist}
\title{Historical/Contemporary questions}
\usage{
is.contemp(x)

is.hist(x)
}
\arguments{
\item{x}{CharacterVector of question tags}
}
\description{
Check whether a given question tag is historical (v3) or
contemporary (v2).
}
\examples{
is.contemp(c("v2clacfree", "v3clacfree"))
is.hist(c("v2csreprss", "v3csreprss"))

}
