% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_tags.R
\name{is.shared_tag}
\alias{is.shared_tag}
\title{Check if a variable exists for both contemporary and historical}
\usage{
is.shared_tag(tags, ttable)
}
\arguments{
\item{tags}{CharacterVector of v2/v3 tag names.}

\item{ttable}{Translation table listing all contemporary (v2) and
historical (v3) variable tag names (typically our
question_table).}
}
\value{
LogicalVector
}
\description{
\code{is.shared_tag} takes a vector of variable tag names and returns
a logical vector indicating which variables exist in both
historical and contemporary.
}
\examples{
vars <- c("v2clacfree", "v3clacfree", "v3elage", "v2elpdcamp")
ttable <- data.frame(name = vars, stringsAsFactors = FALSE)

is.shared_tag(vars, ttable)

}
