% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_date.R
\name{is_vignette}
\alias{is_vignette}
\title{Vignette ratings}
\usage{
is_vignette(x, na.rm = T)
}
\arguments{
\item{x}{Character vector of country-dates}

\item{na.rm}{Whether to set \code{NA} values to false.}
}
\value{
Logical vector
}
\description{
Indicates whether a country-date represents a vignette.
}
\details{
Vignettes are identified by unique constructed from the
vignette type, threshold, and version.
}
\examples{
is_vignette(c("AFG 1900-01-01", "historical_1_v2", "pilot_1_v1"))

}
\seealso{
Other country-date functions: 
\code{\link{is_country_date}()},
\code{\link{sort_text_id}()}
}
\concept{country-date functions}
