% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codemmu.R
\name{kegg_level}
\alias{kegg_level}
\title{diff kegg level metaProfiler}
\usage{
kegg_level(n, level = 2)
}
\arguments{
\item{n}{metaProfiler}

\item{level}{A number of kegg level}
}
\value{
metaProfiler.
}
\description{
diff kegg level metaProfiler
}
\examples{
load(system.file('data', 'Tes.Rdata',package = 'pathways'))
kegg_level(n)
}
