% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_error}
\alias{log_error}
\title{Log an ETL error message}
\usage{
log_error(message)
}
\arguments{
\item{message}{Message to write.}
}
\value{
No meaningful return value.
}
\description{
Writes an error ETL message to the standard error stream.
}
