% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_info}
\alias{log_info}
\title{Log an informational ETL message}
\usage{
log_info(message)
}
\arguments{
\item{message}{Message to write.}
}
\value{
No meaningful return value.
}
\description{
Writes an informational ETL message to the standard output stream.
}
