% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_warn}
\alias{log_warn}
\title{Log an ETL warning message}
\usage{
log_warn(message)
}
\arguments{
\item{message}{Message to write.}
}
\value{
No meaningful return value.
}
\description{
Writes a warning ETL message to the standard output stream.
}
