\name{makeS3Bucket}
\alias{makeS3Bucket}
\title{AWS Support Function: Creates an S3 Bucket...}
\usage{makeS3Bucket(bucketName)}
\description{AWS Support Function: Creates an S3 Bucket}
\details{Creates an S3 bucket. If the bucket already exists, no warning is returned.}
\author{James "JD" Long}
\arguments{\item{bucketName}{string of the name of the bucket to be created}}
