% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva.R
\name{med_dispersao}
\alias{med_dispersao}
\title{Medidas de Dispersão e Escala}
\usage{
med_dispersao(x, decimais = 2)
}
\arguments{
\item{x}{Vetor numérico.}

\item{decimais}{Casas decimais (padrão 2).}
}
\description{
Medidas de Dispersão e Escala
}
