% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runmimix.R
\docType{package}
\name{mimix}
\alias{mimix}
\title{mimix}
\usage{
mimix(
  data,
  covar = NULL,
  depvar,
  treatvar,
  idvar,
  timevar,
  M = 1,
  reference = NULL,
  method = NULL,
  seed = 101,
  prior = "jeffreys",
  burnin = 1000,
  bbetween = NULL,
  methodvar = NULL,
  referencevar = NULL,
  delta = NULL,
  dlag = NULL,
  K0 = 1,
  K1 = 1
)
}
\arguments{
\item{data}{Dataset in wide format}

\item{covar}{Covariates - may include the baseline value of depvar. Must be complete (no missing values).}

\item{depvar}{Dependent (outcome) variable}

\item{treatvar}{Treatment group, coded 1,2,..}

\item{idvar}{Participant id}

\item{timevar}{Time point for repeated measure}

\item{M}{Number of imputations to be created}

\item{reference}{Reference group for J2R, CIR, CR methods}

\item{method}{Reference-based imputation method: must be ...}

\item{seed}{Seed value. Specify this so that a new run of the command will give the same imputed values.}

\item{prior}{Prior for the variance-covariance matrix when fitting multivariate normal distributions. Jeffreys (default), uniform or ridge}

\item{burnin}{Number of burn-in iterations when fitting multivariate normal distributions.}

\item{bbetween}{Number of iterations between imputed data sets when fitting multivariate normal distributions.}

\item{methodvar}{vector designating variable in data-set specifying individual method}

\item{referencevar}{vector designating variable in data-set specifying reference group corresponding to individual method}

\item{delta}{vector of delta values to add onto imputed values (non-mandatory) (a's in Rogers paper),length as number of time points}

\item{dlag}{vector of delta values to add onto imputed values (non-mandatory) (b's in Rogers paper),length as number of time points}

\item{K0}{Causal constant for use with Causal method}

\item{K1}{exponential decaying Causal constant for use with Causal method}
}
\value{
impdataset the M imputed data-sets appended to the "missing values" data-set in wide format
}
\description{
main wrapper for running mimix

similar to the Stata mimix function
}
\details{
This is based on Suzie Cro's Stata program

sets up a summary table based on missing data pattern- mg  mimix_group

reflects the pattern and treatment group configuration of the raw data

then acts as a looping mechanism, norm2 is used as MCMC multivariate normal

The mimix package contains the functions preprodata and preproIndivdata to
process long longitudinal data into wide data format

pass2Loop performs 2nd pass after interims found by MAR

Also the function Addelta to add delta adjustment to the imputed estimates
}
\examples{
\dontrun{
 mimix("asthma",c("base"),"fev","treat","id","time",10,1,"Causal",101,"jeffreys",1000,NULL,NULL,NULL,c(0.5,0.5,1,1),c(1,1,1,1),1,0.6)
}
}
