\name{modalvalue}
\alias{modalvalue}
\title{
  Calculate the mode of a vector, array or list.
}
\description{
  
   The mode is the most common or modal value of a list.  This function
   calculates the mode of a vector, array or list (lists are
   flattened).
  
}
\usage{
modalvalue(x, na.rm=FALSE)
}
\arguments{
  
  \item{x}{A vector, array or list.}

  \item{na.rm}{Whether to remove missing values.}
}
\details{

  Rattle needed this function, and R does not provide one. This version
  came from the R Wiki.

}
\references{R Wiki: \url{http://wiki.r-project.org/rwiki/doku.php?id=tips:stats-basic:modalvalue}}
\keyword{hplot}
