% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoElnet_function.R, R/ParetoFunction.R
\name{myFM}
\alias{myFM}
\title{myFM}
\usage{
myFM(x)

myFM(x)
}
\arguments{
\item{x}{Input predictor weight vector}
}
\value{
f Criterion vector

f Criterion vector
}
\description{
Supporting function, defines criterion space

Supporting function, defines criterion space
}
