% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoElnet_function.R
\name{myT_Elnet}
\alias{myT_Elnet}
\title{myT_Elnet}
\usage{
myT_Elnet(x_t)
}
\arguments{
\item{x_t}{Temporary input weight vector}
}
\value{
f Temporary criterion space
}
\description{
Support function, define criterion space for intermediate step in NBI()
}
