% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.cualquiera.R
\encoding{UTF-8}
\name{NA.cualquiera}
\alias{NA.cualquiera}
\title{Sustituir los NA de un vector por cualquier valor deseado}
\usage{
NA.cualquiera(x, reemplazo)
}
\arguments{
\item{x}{Columna o vector numérico o tipo caracter.}

\item{reemplazo}{Cualquier valor deseado que ser? usado para sustituir los NA.}
}
\value{
Na por cualquier valor en un vector o columna.
}
\description{
Sustituir los NA de un vector por cualquier valor deseado, puede ser numérico o caracter.
}
\examples{
data(iris)
#Introducimos NAs aleatorios en el dataframe
iris$Sepal.Width[c(1,3,5,7)] <- NA
head(iris)
#Ahora vamos a sustituir los NA por cierto valor en especifico, en este caso por 22.22.
iris<-tlamatini::NA.cualquiera(iris,c(22.22))
head(iris)
}
