% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.quitarEspecial.R
\encoding{UTF-8}
\name{NA.quitarEspecial}
\alias{NA.quitarEspecial}
\title{Reemplazar caracteres especiales por NA}
\usage{
NA.quitarEspecial(df, vec)
}
\arguments{
\item{df}{dataframe}

\item{vec}{vector que contiene los caracteres especiales que desea sustituir por NA}
}
\value{
Vector sin caracteres especiales.
}
\description{
Reemplaza los caracteres especiales en el dataframe por NA.
}
\examples{
data(iris)
iris[2,2]<-"@"
iris[3,2]<-"???"
head(iris)
#quitar los caracteres especiales
iris2<-NA.quitarEspecial(iris,c("@","???"))
head(iris2)
}
