% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{normalize_qids}
\alias{normalize_qids}
\title{Normalize historical \code{question_id}s}
\usage{
normalize_qids(ids, ttable)
}
\arguments{
\item{ids}{NumericVector of \code{question_id}s.}

\item{ttable}{\code{question_id} translation table (\emph{e.g.},
the question table).}
}
\description{
Given a vector of \code{question_id}s, \code{normalize_qids}
replaces the historical \code{question_id}s with the analogous
contemporary \code{question_id}s where there's a match based on
\code{ttable}.
}
\details{
\code{normalize_qids} is meant to
work directly with the question table. Thus, \code{ttable}
requires two columns: \code{name} and \code{question_id}. We
determine whether an historical \code{question_id} has a
matching contemporary \code{question_id} by checking the root
tag --- the portion of the tag after removing the \code{v\\d}
suffix.
}
\examples{
ttable <- data.frame(question_id = 1:3,
                     name = c("v2clacfree", "v3clacfree", "v3strenadm"),
                     stringsAsFactors = FALSE)

normalize_qids(1:3, ttable)

}
