% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.ntp.R
\name{NTP}
\alias{NTP}
\title{NTP}
\usage{
NTP(eset, sigList, out.dir, output.name, dist.selection, norm.method, nresmpl, rnd.seed, mc.cores)
}
\arguments{
\item{eset}{expressio}

\item{sigList}{gene signiture list}

\item{out.dir}{output directory}

\item{dist.selection}{calculating distance, a character, either c("correlation" or "cosine").}

\item{norm.method}{normalization method, either c("none", "row.std")}

\item{nresmpl}{an integer, number of permutations for \eqn{p}-value}

\item{rnd.seed}{Seed of the random number generator.}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
Nearest Template Prediction (NTP) result
}
\description{
A wrapper function to NTP
}
\details{
Nearest Template Prediction (NTP) based on predefined class templates.
}
\references{
Hoshida, Y. (2010). Nearest Template Prediction: A Single-Sample-Based Flexible Class Prediction with Confidence Assessment. PLoS ONE 5, e15543.
}
