% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.GLMM2.R
\encoding{UTF-8}
\name{optimizeGLMM2}
\alias{optimizeGLMM2}
\title{Probar optimizadores en modelo glmmTMB}
\usage{
optimizeGLMM2(formula, data)
}
\arguments{
\item{formula}{Formula del modelo que queremos ajustar, incluyendo el componente aleatorio si es que hay. Por ejemplo: "y ~ x+z"}

\item{data}{Base de datos que se usa para ajustar el modelo.}
}
\value{
Lista con los diferentes optimizadores y su ajuste. Si es FALSE, el modelo no converge.Si es TRUE, podemos ajustar ese modelo con ese optimizador.
}
\description{
Esta es una función que permite ajustar GLMM usando una fomula y tus datos probando con diferentes optimizadores cuando existen problemas de convergencia. A veces al ajustar modelos pueden arrojar errores. Por ejemplo, una advertencia ‘iteration limit reached without convergence’, se soluciona aumentando el número de iteraciones: 'glmmTMBControl(optCtrl=list(iter.max=1e3,eval.max=1e3))'. Vea la ayuda de la función glmmTMBControl para mas detalles.
}
\examples{
# do not run
#datos <- ChickWeight
#library(glmmTMB)
#modelo <- glmmTMB(weight ~ Diet +(1|Chick), family=gaussian("log"), data = datos)
# optimizeGLMM2(formula = p_azucar ~ tratamiento * dia, data = datos)
}
