% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ord}
\alias{ord}
\title{Ordinal scale transformation}
\usage{
ord(z, gamma_mu)
}
\arguments{
\item{z}{Extracted Z parameter as a matrix}

\item{gamma_mu}{Extracted gamma_mu as a matrix}
}
\value{
Integer matrix with the same dimensions as our Z matrix
}
\description{
Transforms latent trait estimates into the original ordinal scale.
}
