% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{ParDepCalPlots}
\alias{ParDepCalPlots}
\title{Function automatically builds partial dependence calibration plots for model evaluation}
\usage{
ParDepCalPlots(data, PredColName = "PredictedValues",
  ActColName = "ActualValues", IndepVar = "Independent_Variable_Name",
  type = "calibration", bucket = 0.05,
  Background = "lightsteelblue1", Borders = "navyblue",
  FactLevels = 10, Function = function(x) mean(x, na.rm = TRUE))
}
\arguments{
\item{data}{Data containing predicted values and actual values for comparison}

\item{PredColName}{String representation of the column name with predicted values from model}

\item{ActColName}{String representation of the column name with actual values from model}

\item{IndepVar}{String representation of the column name with the independent variable of choice}

\item{type}{Calibration or boxplot - calibration aggregated data based on summary statistic; boxplot shows variation}

\item{bucket}{Number of buckets to partition the space on (0,1) for evaluation}

\item{FactLevels}{The number of levels to show on the chart (1. Levels are chosen based on frequency; 2. all other levels grouped and labeled as "Other")}

\item{Function}{Supply the function you wish to use for aggregation.}
}
\value{
Partial dependence calibration plot or boxplot
}
\description{
This function automatically builds partial dependence calibration plots and partial dependence calibration boxplots for model evaluation using regression, quantile regression, and binary and multinomial classification
}
\examples{
aa <- data.table(target = runif(10000))
aa[, x1 := qnorm(target)]
aa[, x2 := runif(10000)]
aa[, Independent_Variable := log(pnorm(0.75 * x1 + sqrt(1-0.75^2) * qnorm(x2)))]
aa[, ':=' (x1 = NULL, x2 = NULL)]
h2o.init()
data <- as.h2o(aa)
model <- h2o.gbm(x = 2, y = 1, training_frame = data)
preds1 <- as.data.table(h2o.predict(model, data))
preds2 <- cbind(preds1, aa)
data <- preds2
output <- ParDepCalPlots(preds2,
                         PredColName = "predict",
                         ActColName  = "target",
                         IndepVar    = "Independent_Variable",
                         type        = "calibration", # c("calibration","boxplot"),
                         bucket      = 0.05,
                         Background  = "lightsteelblue1",
                         Borders     = "navyblue",
                         FactLevels  = 10,
                         Function    = function(x) mean(x, na.rm = TRUE))
}
\author{
Adrian Antico at RemixInstitute.com
}
