% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoAdj.R
\name{ParetoAdj}
\alias{ParetoAdj}
\title{Pareto_Adj function}
\usage{
ParetoAdj(Ncal, wpred, Rperf_cal, Rrace_cal, prop, sr)
}
\arguments{
\item{Ncal}{Calibration sample size}

\item{wpred}{Matrix of predictor weights, where each row displays a Pareto-optimal solution and each column displays a weight for each predictor in that solution}

\item{Rperf_cal}{Vector of calibration sample job performance validity}

\item{Rrace_cal}{Vector of calibration sample race bivariate correlation (i.e., correlation between race dummy variable (0-minority, 1-majority) and predictor composite score)}

\item{prop}{Proportion of minority applicants in full applicant pool}

\item{sr}{Selection ratio}
}
\value{
A data frame, each row shows a Pareto-optimal solution prior to and after the adjustment using the Pareto-optimal shrinkage formula

\describe{
  \item{Rperf_cal, Rperf_adj}{Job performance validity prior to and after adjustment, respectively}
  \item{Rrace_cal, Rrace_adj}{Race bivariate correlation prior to and after adjustment, respectively}
  \item{AIratio_cal, AIraio_adj}{Adverse impact ratio prior to and after adjustment, respectively}
  \item{w_perf, w_race}{The extent to which, respectively, job performance validity and race bivariate correlation is optimized in a given Pareto-optimal solution}
  }
}
\description{
Estimate shrunken Pareto-optimal solution based on Pareto-optimal shrinkage formulas
}
\examples{
# Example corresponds to Song, Tang, Newman, & Wee (2023), Supplemental Material 4
# Song, Q. C., Tang, C., Newman, D. A., & Wee, S. (2023). Adverse impact reduction
# and job performance optimization via Pareto-optimal weighting: A shrinkage formula
# and regularization technique using machine learning. Journal of Applied
# Psychology. doi:10.1037/apl0001085.
# (1) Calibration sample size
Ncal <- 40
# (2) Predictor weights for each Pareto solution.
# Rows: Pareto solutions; Columns: Predictors
wpred <- matrix(c(0,    0,    0,    0,    1,
                  0,    0,    0, 0.07, 0.93,
                  0,    0,    0, 0.13, 0.87,
                  0,    0,    0, 0.18, 0.82,
                  0,    0,    0, 0.23, 0.77,
                  0,    0,    0, 0.28, 0.72,
                  0,    0,    0, 0.32, 0.68,
                  0,    0,    0, 0.37, 0.63,
                  0,    0,    0, 0.41, 0.59,
                  0, 0.01,    0, 0.45, 0.55,
                  0, 0.04,    0, 0.44, 0.52,
                  0, 0.07,    0, 0.43,  0.5,
                  0,  0.1,    0, 0.42, 0.48,
                  0, 0.14,    0, 0.41, 0.46,
                  0, 0.17,    0,  0.4, 0.43,
                  0, 0.21,    0, 0.38, 0.41,
                  0, 0.25,    0, 0.37, 0.38,
                  0,  0.3,    0, 0.35, 0.35,
                  0, 0.33, 0.03, 0.34,  0.3,
                  0, 0.37, 0.08, 0.31, 0.25,
                  0, 0.42, 0.15, 0.27, 0.15), ncol = 5, nrow = 21)
# (3) Vector of calibration sample job performance validity
Rperf_cal = c(.20, .24, .27, .30, .33, .36, .39, .42, .45, .48, .51, .54,
.57, .60, .63, .66, .69, .72, .74, .76, .78)
# (4) Vector of calibration sample race bivariate correlation
# (i.e., correlation between race dummy variable (0-minority, 1-majority)
# and predictor composite score)
Rrace_cal = c(-.12, -.11, -.10, -.10, -.09, -.08, -.07, -.06, -.05, -.03,
-.02, .00, .01, .03, .05, .07, .09, .12, .15, .19, .24)
# (5) proportion of minority
prop <- 1/6
# (6) selection ratio
sr <- .15
# Estimate shrunken Pareto-optimal solution
ParetoAdj(Ncal = Ncal, wpred = wpred, Rperf_cal = Rperf_cal,
Rrace_cal = Rrace_cal, prop = prop, sr = sr)
}
