% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blank.R
\name{plot_blank}
\alias{plot_blank}
\title{Funcion complementaria de auto mfrow tomada de la paqueteria mjcgraphics}
\usage{
plot_blank(main = NULL, sub = NULL, message = NULL, box = FALSE)
}
\arguments{
\item{main}{you can optionally write a main title, and a sub-title}

\item{sub}{you can optionally write a main title, and a sub-title}

\item{message}{message to write in the middle of the plotting device}

\item{box}{logical: draw a box around the plot?}
}
\description{
automaticamente elige cuantas columnas y filas se usaran para
mostrar las graficas \code{par(mfrow)}. creada por Mark Cowley, 3 June 2006
}
\keyword{internal}
