% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ds.R
\name{plot_ds}
\alias{plot_ds}
\title{plot_ds}
\usage{
plot_ds(dataset, x = NULL)
}
\arguments{
\item{dataset}{str. Name of the dataset to plot}

\item{x}{char. Name of the column of interest of the dataset to plot}
}
\value{
plot
}
\description{
It plots the data depending on the dataset used
}
\examples{
\dontrun{
plot_ds(dataset=canada_ab,x=province)
}

}
