% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva.R
\name{quartis}
\alias{quartis}
\title{Quartis da Distribuição}
\usage{
quartis(x, decimais = 2)
}
\arguments{
\item{x}{Vetor numérico.}

\item{decimais}{Casas decimais (padrão 2).}
}
\description{
Quartis da Distribuição
}
