\name{read_asc}
\alias{read_asc}
\title{Reads a triple-s XML (...}
\usage{read_asc(filename)}
\description{Reads a triple-s XML (.asc) data file, as specified by the triple-s XML standard}
\details{This function reads and parses a .asc XML data file.
The .sss standard defines a standard survey structure}
\keyword{XML}
\arguments{\item{filename}{Name of .asc file containing the survey metadata}}
\examples{# Not executed
# read("sample.sss")}
