\name{ReadSofiaOpt}
\alias{ReadSofiaOpt}
\title{Read results from an SOFIA optimization experiment }
\description{The optimization within \code{\link{SofiaOpt}} produces files that can be used to restart or monitor an optimization experiment. This function reads these files.}
\usage{ReadSofiaOpt(files, combine = TRUE, ...)}
\arguments{
  \item{files}{vector of file names}
  \item{combine}{combine several files in a single file?}
  \item{\dots}{further arguments (not used)}
}
\details{No details.}

\references{No reference.	}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}



\seealso{\code{\link{SofiaOpt}}}

