% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regr_chickw.R
\name{regr_chickw}
\alias{regr_chickw}
\title{regr_chickw}
\usage{
regr_chickw(z)
}
\arguments{
\item{z}{character. Name of a column from the dataset 'regr_chickw' to regress}
}
\value{
table, str
}
\description{
Regress a variable z from the dataset 'regr_chickw' on the variable 'weight'.
If the intercept value is non-significant, it computes again the regression without the intercept. It also returns if the regressed variable is significant or not
}
\examples{
\dontrun{
regr_chickw(z=Time)
}

}
