% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{RemixTheme}
\alias{RemixTheme}
\title{RemixTheme function is a ggplot theme generator for ggplots}
\usage{
RemixTheme()
}
\value{
An object to pass along to ggplot objects following the "+" sign
}
\description{
This function adds the Remix Theme to ggplots
}
\examples{
p <- ggplot(data, aes(x = weeks, y = quantity)) + geom_line()
p <- p + RemixTheme()
p
}
\author{
DougVegas at RemixInstitute.com
}
