% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_DHARMa.R
\encoding{UTF-8}
\name{resid_DHARMa}
\alias{resid_DHARMa}
\title{Grafico de residuales de un modelo LM, LMM, GLM, GLMM, GAM, etc}
\usage{
resid_DHARMa(Modelo, nsim = NULL)
}
\arguments{
\item{Modelo}{Un modelo LM, GLM, GLMM, GAM.}

\item{nsim}{Número de simulaciones. Cuanto menor sea el número, mayor será el error estocástico en los residuales.
Además, para n muy pequeño, los artefactos de discretización pueden influir en las pruebas. Por defecto
es 250, que es un valor relativamente seguro. Puede considerar aumentar a 1000 para estabilizar
los valores simulados.}
}
\value{
Grafica los residuales de un modelo.
}
\description{
Muestra el ajuste del modelo LM, GLM, GLMM, GAM, etc. Esta función crea residuos escalados
mediante la simulación del modelo ajustado, 250 simulaciones por default. Los residuos pueden ser extraídos con residuals.DHARMa. Ver
testResiduals para una visión general de las pruebas de residuos, plot.DHARMa para una visión general de
los gráficos disponibles. Esta función fue creada para explorar el ajuste, normalidad y homogeneidad de
varianzas entre los grupos.
}
\examples{
data(warpbreaks)
modelo <- glm(breaks ~ wool + tension, family= poisson("log"), data= warpbreaks)
resid_DHARMa(modelo)

### tambien se puede usar en GLMMs:
datos <- datasets::ChickWeight
library(lme4)
modelo <- glmer(weight ~ Diet +(1|Chick), family=gaussian("log"), data = datos)
resid_DHARMa2(modelo)
}
\references{
\itemize{
\item Ali MM, Sharma SC (1996) Robustness to nonnormality of regression F-tests. J Econom71, 175–205.

\item Lumley T, Diehr P, Emerson S, Chen L (2002) The importance of the normality assumption in
large public health data sets. Annu Rev Public Health23, 151–169.
}
}
