% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_plot.R
\encoding{UTF-8}
\name{resid_plot}
\alias{resid_plot}
\title{Grafica de residuales de un modelo. Version 2}
\usage{
resid_plot(Modelo)
}
\arguments{
\item{Modelo}{Un modelo LM, GLM o GLMM.}
}
\value{
Esta función fue creada para explorar el ajuste, normalidad y homogeneidad de varianzas entre
los grupos.
}
\description{
Muestra el ajuste del modelo LM, GLM o GLMM. Esta función permite la exploración
visual de los residuales de un modelo.
}
\examples{
data(iris)
modelo <- glm(Petal.Width ~ Petal.Length, family = gaussian("log"), data=iris)
resid_plot(modelo)
}
