% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{robust.theta}
\alias{robust.theta}
\title{Model: Rasch, 1PL, 2PL, MIRT, GRM, MGRM
Weight: equal, bisquare, Huber
Residual: information, standardized, MSR, dual-residual
Estimation procedure: MLE, MAP, EAP
Standard error: asymptotic, sandwich
dimen: specify number of dimensions for multidimensional data
Ability Estimation Function Using Robust Estimation (GRM)}
\usage{
robust.theta(
  dat,
  ipars,
  model = "2PL",
  D = 1.7,
  dimen = NULL,
  weight.type = "equal",
  tuning.par = NULL,
  residual = "standardized",
  est.type = "MLE",
  prior = c(0, 1),
  eap.quad.pts = seq(-4, 4, length.out = 41),
  iter = 30,
  tol = 0.01
)
}
\arguments{
\item{dat}{A \eqn{N \times J} matrix of polytomously-scored data (e.g., Likert-type) for \emph{N} subjects and \emph{J} items. Indexing begins at 0.}

\item{ipars}{A matrix of item parameters, whose structure depends on the model.}

\item{model}{Character string specifying the IRT model. See `item.prob` for supported models.}

\item{D}{A positive scaling constant used for scaling the normal ogive model. Defaults to 1.7; alternatively is often set to 1.0.}

\item{weight.type}{Type of weighting function to be used: "equal", "Huber", or "bisquare".}

\item{tuning.par}{Optional tuning parameter for Huber or bisquare weights.}

\item{est.type}{Type of estimation to be used: "MLE", "MAP", or "EAP".}

\item{prior}{Numeric vector giving the mean and variance of the normal prior for MAP and EAP estimation types.}

\item{eap.quad.pts}{A numeric vector of quadrature points \eqn{\theta_q} used when computing EAP standard errors. These are typically q equally spaced values across the latent‑trait range. Only used when "EAP" is included in \code{est.type}.}

\item{theta}{A numeric vector or matrix of latent trait values. For unidimensional models, a numeric vector of length \eqn{N}. For multidimensional models (MIRT, MGRM), an \eqn{N \times L} matrix.}
}
\value{
theta Ability estimates for \emph{N} subjects. NAs replace values that did not converge to any value. Estimates that converged to values less than -3.0 were replaced with -3.0, while estimates that converged to values greater than 3.0 were replaced with 3.0.

convergence Indicators of convergence for \emph{N} subjects: a “0” indicates the value converged, while a “1” indicates the maximum likelihood estimation did not converge to any value.

standard.error Standard errors of the theta estimates for \emph{N} subjects, given by the square root of the reciprocal of the Fisher information. NAs replace nonconverging values.

theta.progression A matrix with rows corresponding to each subject and columns corresponding to the number of iterations supplied to the input. Each column provides the updated theta estimate at each iteration of the Newton-Raphson algorithm until the change in log-likelihood for that subject reaches the cutoff value or the value is nonconverged (reaches infinite values).

residual A \eqn{J \times N \times p} array containing residuals corresponding to the ability estimate for \emph{N} subjects respective to the \emph{J} test items at each iteration until convergence within maximum \emph{p} iterations, nonconvergence, or singular matrix is reached.
}
\description{
Ability Estimation Function Using Robust Estimation
}
\details{
The goal of robust estimation is to downweigh potentially aberrant responses to lessen their impact on the estimation of \eqn{\theta_i}. Robust estimates resist the harmful effects of response disturbances and tend to be less biased estimates of true ability than maximum likelihood estimates.
              Under a given item response model, let the probability of examinee \eqn{i} endorsing exactly category \eqn{k} on item \eqn{j} be denoted \eqn{P_{ijk}}.
              The contribution of item \emph{j} to the overall log-likelihood for one subject is weighted with a weight \eqn{\omega(r_{ij})} as a function of a residual \eqn{r_{ij}} for the item:
              \deqn{\sum^J_{j=1} \omega(r_{ij}) \sum^K_{k=1} u_{ijk}\text{log}P_{ijk} = 0 }
              \eqn{u_{jk}} is an indicator function: \deqn{u_{ijk} = \begin{cases}
                                                           1 & \text{if } X_{ij} = k; \\
                                                           0 & \text{otherwise}.
                                                           \end{cases} }
              The residual, which measures the inconsistency of a response from the subject's assumed response model, is \deqn{r_{ij} = \frac{1}{\sigma_{X_{ij}}}\left[X_{ij} - E(X_{ij}|\hat{\boldsymbol{\theta}}_i)\right]}.
              The difference in fit is determined between the observed response \eqn{X_{ij}} and expected score \eqn{E(X_{ij}|\hat{\boldsymbol{\theta}}_i) = \sum_{k=1}^KkP_{jk}(\hat{\boldsymbol{\theta}}_i)}, and scaled by the variance \eqn{\sigma_{X_{ij}}^2 = \sum_{k=1}^K (X_{ijk}-E[X_{ij}|\hat{\theta}_i])^2P_{jk}(\hat{\theta}_i).}
              Two types of weight functions are used: Tukey's bisquare weighting function (Mosteller & Tukey, 1977)
                \deqn{\omega(r_{ij})=\begin{cases}[1-(r_{ij}/B)^2]^2, & \text{if} |r_{ij}|\leq B.\\0, & \text{if} |r_{ij}|>B.\end{cases}}
              and the Huber weighting function (Huber, 1981)
                \deqn{\omega(r_{ij})=\begin{cases}1, & \text{if} |r_{ij}|\leq H.\\H/|r_{ij}|, & \text{if} |r_{ij}|>H.\end{cases}}
              Both functions are effective in estimating more accurate scores with aberrant data, although the bisquare weight function may lead to nonconvergence when using data containing a high proportion of incorrect responses (Schuster & Yuan, 2011).
}
\references{
Embretson, S. E., & Reise, S. P. (2000). \emph{Item response theory for psychologists.} Mahwah, N.J: L. Erlbaum Associates.

Huber, P. (1981) \emph{Robust Statistics}. Wiley, New York. https://doi.org/10.1002/0471725250

Mosteller, F., & Tukey, J. W. (1977). \emph{Data Analysis and Regression: A Second Course in Statistics}. Reading, MA: Addison-Wesley Pub Co.

Samejima, F. (1969). Estimation of latent ability using a response pattern of graded scores. \emph{Psychometrika Monograph Supplement, 34} (4, Pt. 2), 100–100.

Schuster, C., & Yuan, K.-H. (2011). Robust Estimation of Latent Ability in Item Response Models. \emph{Journal of Educational and Behavioral Statistics}, 36(6), 720–735. https://doi.org/10.3102/1076998610396890
}
