% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fredi_sv.R
\name{run_fredi_sv}
\alias{run_fredi_sv}
\title{Calculate climate change impacts on socially vulnerable (SV) populations throughout the 21st century for available sectors}
\usage{
run_fredi_sv(
  sector = NULL,
  driverInput = NULL,
  popInput = NULL,
  silent = TRUE,
  .testing = FALSE
)
}
\arguments{
\item{sector}{A character string indicating the sector for which the FrEDI SV module should calculate impacts (see \code{\link[=get_sv_sectorInfo]{get_sv_sectorInfo()}} for a list of available sectors).}

\item{driverInput}{A data frame of up to four custom scenarios for drivers (temperature or global mean sea level rise). \code{driverInput} requires a data frame with columns of \code{"year"} and \code{"scenario"}. The data frame must also include a third column: \code{"temp_C"} for temperature-driven sectors (containing temperature values in degrees Celsius of warming for the contiguous U.S.) or \code{"slr_cm"} for sea level rise (SLR)-driven sectors (containing values for global mean sea level rise in centimeters). Run \code{get_sv_sectorInfo(gcmOnly=TRUE)} to see temperature-driven sectors in the SV module and \code{get_sv_sectorInfo(slrOnly=TRUE)} to see SLR-driven scenarios. Users can also pass a data frame with all four columns (\code{"year"}, \code{"scenario"}, \code{"temp_C"}, and \code{"slr_cm"}), in which case \code{\link[=run_fredi_sv]{run_fredi_sv()}} determines whether to use the \code{"temp_C"} or \code{"slr_cm"} column as the driver trajectory based on the specified sector. Driver inputs for all scenarios should start in the year 2000 or earlier. All scenarios must include at least two non-missing values  (especially values before or at 2000 and at or after 2100). If any required columns are missing, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will use the default temperature or sea level rise scenario from \code{\link[=run_fredi]{run_fredi()}}. If the data frame passed to \code{driverInput} has more than four unique scenarios, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will only run the first four scenarios.}

\item{popInput}{The input population scenario requires a data frame object with a single scenario of state-level population values.
\itemize{
\item The population scenario must have five columns with names \code{"year"}, \code{"region"}, \code{"state"}, \code{"postal"}, and \code{"state_pop"} containing the year, the NCA region name, the state name, the postal code abbreviation (e.g., "ME" for "Maine") for the state, and the state population, respectively.
\item \code{popInput} only accepts a a single scenario, in contrast to \code{driverInput}. In other words, \code{\link[=run_fredi_sv]{run_fredi_sv()}} uses the same population scenario for any and all driver scenarios passed to \code{driverInput}.
\item If the user does not specify an input scenario for population (i.e., \code{popInput = NULL}, \code{\link[=run_fredi_sv]{run_fredi_sv()}} uses a default population scenario.
\item Population inputs must have at least one non-missing value in 2010 or earlier and at least one non-missing value in or after the final analysis year (2100).
\item Population values must be greater than or equal to zero.
}}

\item{silent}{A logical (\code{TRUE/FALSE}) value indicating the level of messaging desired by the user (defaults to \code{silent=TRUE}).}
}
\value{
The output of \code{\link[=run_fredi_sv]{run_fredi_sv()}} is an R data frame object containing average annual physical impacts for socially vulnerable groups, at the NCA region level and five-year increments.
}
\description{
\code{run_fredi_sv} allows users to project annual average climate change impacts throughout the 21st century (2010-2100) for socially vulnerable (SV) populations for available sectors. Users can run \code{\link[=run_fredi_sv]{run_fredi_sv()}} for individual sectors to generate annual physical impacts for SV populations. \code{\link[=run_fredi_sv]{run_fredi_sv()}} can be run with default population and climate (temperature and sea level rise trajectories) or provide custom trajectories. The output of \code{\link[=run_fredi_sv]{run_fredi_sv()}} is an R data frame object containing annual average physical impacts at five-year increments for the period 2010 to 2100. Users have the option to write outputs to Excel files that provide additional visualization of SV outputs.
}
\details{
\code{\link[=run_fredi_sv]{run_fredi_sv()}} projects annual climate change impacts for socially vulnerable (SV) populations throughout the 21st century (2010-2100) for available sectors, using default or user-specified population, temperature, and sea level rise (SLR) trajectories. \code{\link[=run_fredi_sv]{run_fredi_sv()}} is the main function for the FrEDI Social Vulnerability (SV) module in the \link{FrEDI} R package, described elsewhere (See \url{https://epa.gov/cira/FrEDI} for more information). The SV module extends the \link{FrEDI} framework to socially vulnerable populations using data underlying a 2021 U.S. Environmental Protection Agency (EPA) report on \href{https://www.epa.gov/cira/social-vulnerability-report/}{Climate Change and Social Vulnerability in the United States}.

Users can run \code{\link[=run_fredi_sv]{run_fredi_sv()}} to generate annual physical impacts for SV groups for individual sectors. When running \code{\link[=run_fredi_sv]{run_fredi_sv()}}, users must specify one of the sectors in the SV module; use \code{\link[=get_sv_sectorInfo]{get_sv_sectorInfo()}} for a list of available sectors.

\code{\link[=run_fredi_sv]{run_fredi_sv()}} can be run with default population and climate (temperature and SLR) trajectories or use \code{\link[=run_fredi_sv]{run_fredi_sv()}} to run custom scenarios. Running \code{\link[=run_fredi_sv]{run_fredi_sv()}} with custom climate scenarios requires passing a data frame of scenarios to the \code{driverInput} argument. \code{\link[=run_fredi_sv]{run_fredi_sv()}} can also be run with a custom population scenario by passing a data frame of regional population trajectories to the \code{popInput} argument; unlike climate scenarios, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will only run a single scenario at a time.
\itemize{
\item \code{driverInput} can take a data frame containing up to four custom scenarios for drivers (temperature or global mean sea level rise). \code{driverInput} requires a data frame with columns of \code{"year"} and \code{"scenario"}. The data frame must also include a third column: \code{"temp_C"} for temperature-driven sectors (containing temperature values in degrees Celsius of warming for the contiguous U.S.) or \code{"slr_cm"} for sea level rise (SLR)-driven sectors (containing values for global mean sea level rise in centimeters). Run \code{get_sv_sectorInfo(gcmOnly = TRUE)} to see temperature-driven sectors in the SV module and \code{get_sv_sectorInfo(slrOnly = TRUE)} to see SLR-driven scenarios. Users can also pass a data frame with all four columns (\code{"year"}, \code{"scenario"}, \code{"temp_C"}, and \code{"slr_cm"}), in which case \code{\link[=run_fredi_sv]{run_fredi_sv()}} determines whether to use the \code{"temp_C"} or \code{"slr_cm"} column as the driver trajectory based on the specified sector. If any required columns are missing, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will use the default temperature or sea level rise scenario from \code{\link[=run_fredi]{run_fredi()}}. If the data frame passed to \code{driverInput} has more than four unique scenarios, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will only run the first four scenarios.
\itemize{
\item Temperature inputs must be temperature change in degrees Celsius for the contiguous U.S. (use \code{\link[=convertTemps]{convertTemps()}} to convert global temperatures to CONUS temperatures before passing to \code{driverInput}) relative to a 1995 baseline (where 1995 is the central year of a 1986-2005 baseline period; values should start at zero in the year 1995).
\item Sea level rise inputs must be in centimeters relative to a 2000 baseline (i.e., values should start at zero in the year 2000). Driver inputs for all scenarios should start in the year 2000 or earlier. All scenarios must include at least two non-missing values  (especially values before or at 2000 and at or after 2100).
}
\item The input population scenario requires a data frame object with a single scenario of state-level population values.
\itemize{
\item The population scenario must have five columns with names \code{"year"}, \code{"region"}, \code{"state"}, \code{"postal"}, and \code{"state_pop"} containing the year, the NCA region name, the state name, the postal code abbreviation (e.g., "ME" for "Maine") for the state, and the state population, respectively.
\item \code{popInput} only accepts a a single scenario, in contrast to \code{driverInput}. In other words, \code{\link[=run_fredi_sv]{run_fredi_sv()}} uses the same population scenario for any and all driver scenarios passed to \code{driverInput}.
\item If the user does not specify an input scenario for population (i.e., \code{popInput = NULL}, \code{\link[=run_fredi_sv]{run_fredi_sv()}} uses a default population scenario.
\item Population inputs must have at least one non-missing value in 2010 or earlier and at least one non-missing value in or after the final analysis year (2100).
\item Population values must be greater than or equal to zero.
The default regional population scenario is drawn from the Integrated Climate and Land Use Scenarios version 2 (ICLUSv2) model (Bierwagen et al, 2010; EPA 2017) under the Median variant projection of United Nations (United Nations, 2015). Note that the FrEDI SV default population scenario differs from the default population scenario used by \code{\link[=run_fredi]{run_fredi()}}.
}
}

The output of \code{\link[=run_fredi_sv]{run_fredi_sv()}} is an R data frame object containing average annual physical impacts for socially vulnerable groups, at the NCA region level and five-year increments.
}
\examples{
### Run SV Module with defaults without specifying sector
df_sv <- run_fredi_sv()

### Return a character vector with the names of all of the sectors in the FrEDI SV Module:
get_sv_sectorInfo()

### Return a data frame of all of the sectors in the FrEDI SV Module (sector names and additional information)
get_sv_sectorInfo(description=T)

### Run SV Module with defaults for "Coastal Properties" without saving
df_sv <- run_fredi_sv(sector="Coastal Properties")

### Run SV Module with defaults for "Extreme Temperature" without saving
df_sv <- run_fredi_sv(sector="Extreme Temperature")

### Load temperature scenarios
load(gcamScenarios)

### Load population scenario
load(popScenario)

### Run SV Module for "Extreme Temperature" with custom population and temperature scenarios
df_sv <- run_fredi_sv(sector = "Extreme Temperature", driverInput = gcamScenarios, popInput = popScenario)

}
\references{
Bierwagen, B., D. M. Theobald, C. R. Pyke, A. Choate, P. Groth, J. V. Thomas, and P. Morefield. 2010. “National housing and impervious surface scenarios for integrated climate impact assessments.” Proc. Natl. Acad. Sci. 107 (49): 20887–20892. https://doi.org/10.1073/pnas.1002096107.

EPA. 2017. Multi-Model Framework for Quantitative Sectoral Impacts Analysis: A technical report for the Fourth National Climate Assessment. U.S. Environmental Protection Agency, EPA 430-R-17-001.

EPA. 2021. Technical Documentation on the Framework for Evaluating Damages and Impacts (FrEDI). U.S. Environmental Protection Agency, EPA 430-R-21-004. Available at \url{https://epa.gov/cira/FrEDI/}.

EPA. 2021. Climate Change and Social Vulnerability in the United States: A Focus on Six Impacts. U.S. Environmental Protection Agency, EPA 430-R-21-003. Available at \url{https://www.epa.gov/cira/social-vulnerability-report/}.

United Nations. 2015. World population prospects: The 2015 revision. New York: United Nations, Department of Economic and Social Affairs, Population Division.

U.S. Global Change Research Program. 2015. Scenarios for the National Climate Assessment. Available at \url{https://scenarios.globalchange.gov/regions_nca4}.
}
