% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample calibration data}
\format{
A data frame with 100 rows and 7 columns (5 predictors, criterion, race):
\describe{
  \item{Biodata}{Biodata}
  \item{CogA}{Cognitive ability}
  \item{Con}{Conscientiousness}
  \item{StruInterviw}{Structured interview}
  \item{Integrity}{Integrity test}
  \item{Perf}{Job performance}
  \item{Race_dummy}{Dummy coded race variable; 0-minority, 1-majority}
  ...
}
}
\usage{
sample_data
}
\description{
All columns are numeric and standardized
}
\details{
A sample calibration data used to demonstrate the example usage of cvParetoElnet
}
\keyword{datasets}
