\name{segue-package}
\alias{segue-package}
\title{A segue into parallel processing on Amazon's Web Services.}
\description{A segue into parallel processing on Amazon's Web Services.
A segue into parallel processing on Amazon's Web Services. Includes a
parallel lapply function for the Elastic Map Reduce (EMR) engine. This is
NOT a map reduce project but it uses Hadoop Streaming on Amazon's EMR in
order to get simple parallel computation ala cart. This is NOT a big data
package. If you are CPU bound this might be a useful package.
"We are stuck with technology when what we really want is just stuff that
works." - Douglas Adams}
\details{\tabular{ll}{ Package: \tab segue\cr Type: \tab Package\cr Version: \tab
0.01\cr Date: \tab 2010-12-08\cr License: \tab MIT License \cr LazyLoad:
\tab yes\cr } ~~ An overview of how to use the package, including the most
important ~~ ~~ functions ~~}
\docType{package}
\author{Maintainer: James "JD" Long}
\references{http://code.google.com/p/segue/}
\keyword{package}
