% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sf.R
\name{sf_odbc_driver}
\alias{sf_odbc_driver}
\title{Find Snowflake ODBC Driver Name}
\usage{
sf_odbc_driver()
}
\value{
Character string with driver name
}
\description{
Auto-detects the Snowflake ODBC driver name on the current system.
Can be overridden via the \code{PINS_SF_DRIVER} environment variable.
}
\keyword{internal}
