% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simageData.R
\name{simageData}
\alias{simageData}
\title{Simulates age data}
\usage{
simageData(A, M, FM, f.vul, sd.fvul, s.vul, sd.svul, Popsize, seed,
  fishery = TRUE)
}
\arguments{
\item{A}{= max number of ages classes, M = Average Natural Mortality, FM = Average Fishing Mortality, f.vul = Age first vulnerable to fishery, sd.fvul = sd of age vulnerable to fishery, Fa = Fishing mortality rate for fish of age a, FM = Fishing mortality rate, M = Average natural mortality rate (per year), Na = Total numbers of fish of age a, Ca = The number of fish of age a captured, Popsize = Population size, fishery = The number of fish of age a retained for age sample from fishery (TRUE) and survey (FALSE)}
}
\value{
Simulated age data
}
\description{
User-specified age-structured population model to generate ages from fishery-dependent or independent data such that samples would vary in age composition and abundance
}
