% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlengthVBGF.R
\name{simlengthVBGF}
\alias{simlengthVBGF}
\title{Individual-based model to simulate age:length data from a von Bertalanffy growth function}
\usage{
simlengthVBGF(ages, Linf.mu, Linf.cv, k.mu, k.cv, t0.mu, t0.cv, seed)
}
\arguments{
\item{ages, Linf.mu, Linf.cv, k, k.cv, t0.mu, t0.cv}{= von Bert growth parameters, seed=seed}
}
\value{
Simulated length data
}
\description{
Simulates length data from ages
}
