% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sin_cerosDF.R
\encoding{UTF-8}
\name{sin_cerosDF}
\alias{sin_cerosDF}
\title{Quitar todos los ceros de un dataframe, para quedarnos con observaciones completas}
\usage{
sin_cerosDF(dataframe)
}
\arguments{
\item{dataframe}{Un dataframe variables.}
}
\value{
quitar ceros del vector.
}
\description{
Quitar todos los ceros del dataframe y quedarse solo con las filas con observaciones completas.
}
\examples{
df<-sin_cerosDF(iris)
df
}
