\name{SOfireA-package}
\alias{SOfireA-package}
\alias{SOfireA}
\docType{package}
\title{Satellite Observations for Fire Activity}
}
\description{SOFIA (Satellite Observations for Fire Activity) is an empirical modelling concept to predict burned area based on satellite and climate data. The package implements the basic SOFIA model structure, and functions to optimize and plot SOFIA models.}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SOfireA}
\packageIndices{SOfireA}
}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}

Maintainer: \packageMaintainer{SOfireA}
}

\keyword{ package }


