\name{stopCluster}
\alias{stopCluster}
\title{Stops a running cluster...}
\usage{stopCluster(clusterObject)}
\description{Stops a running cluster}
\details{Stops a running cluster and deletes temporary directories from EC2}
\value{nothing}
\author{James "JD" Long}
\arguments{\item{clusterObject}{a cluster object to stop}}
