% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_ANOVA3F.R
\encoding{UTF-8}
\name{table_ANOVA3F}
\alias{table_ANOVA3F}
\title{Extraer tabla de ANOVA Tipo III usando el estadistico F}
\usage{
table_ANOVA3F(modelo)
}
\arguments{
\item{modelo}{Un modelo GLM,y LMM.}
}
\value{
Tabla ANOVA similar a la que se obtiene con la paqueteria car::Anova.
}
\description{
Extraer tabla de ANOVA Tipo III usando el estadistico F. Extraer la tabla como HTML con formato publicable. Puede ser que algunos modelos no sean soportados y deberían llamarse directamente con la función Anova de la paquetería car.
}
\examples{
data(iris)
modelo <- glm(Petal.Width ~ Petal.Length, family = gaussian("log"), data=iris)
table_ANOVA3F(modelo)
}
