% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_contrasts2.R
\encoding{UTF-8}
\name{table_contrasts2}
\alias{table_contrasts2}
\title{Exportar tabla de contrastes post hoc. Tres variables en interacción}
\usage{
table_contrasts2(cont.emmeans)
}
\arguments{
\item{cont.emmeans}{Contrastes hechos con emmeans de un modelo LM,GLM, GLMM, GAM, LMM, etc.}
}
\value{
Tabla de contrastes.
}
\description{
Extraer los contrastes de un modelo, para generar una tabla en HTML copiable en Word. Esta funcion se usa
cuando tenemos contrastes entre tres grupos, por ejemplo la interacción entre tratamiento x tipo x  sexo.
Para usar esta función es necesario crear un objeto con los contrastes previamente hechos con la
paqueteria emmeans.
}
\examples{
#ejemplo, no correr:
#cont2 <- emmeans(modelo, pairwise ~ Sexo|tratamiento|tipo, adjust="tukey")$contrasts
#table_contrasts2(cont2)
}
