% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_means.R
\encoding{UTF-8}
\name{table_means}
\alias{table_means}
\title{Exportar tabla de medias, SE e intervalos de confianza de un modelo}
\usage{
table_means(cont.emmeans)
}
\arguments{
\item{cont.emmeans}{Medias obtenidas con emmeans de un modelo LM,GLM, GLMM, GAM, LMM, etc.}
}
\value{
Tabla de contrastes con medias, se e IC.
}
\description{
Extraer las medias entre grupos de un modelo, para generar una tabla en HTML copiable en Word.
}
\examples{
#mn2 <- emmeans(m4,pairwise ~ Sexo|tratamiento|tipo, adjust="tukey", type="response")$emmeans
#table_means(mn2)
}
