% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temas_ggplot.R
\name{tema_articulo2}
\alias{tema_articulo2}
\title{Tema con valores predeterminados minimalistas (en mi opinion) adecuados para la publicacion}
\usage{
tema_articulo2(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
Esta función permite usarse como tema predefinido de ggplot para una publiacion científica o para una presentación. Anteriormente esta función estaba con el nombre "my_ggplot_themes".
}
\examples{
library(ggplot2)

d = data.frame(
  x = 1:90,
  y = rnorm(90),
  red = rep(letters[1:3], 30),
  blue = c(rep(1, 30), rep(2, 30), rep(3, 30)))

p <- ggplot(d) +
  geom_point(aes(x = x, y = y)) +
  facet_grid(red ~ blue)
p + tema_articulo()
p + tema_articulo(axis.size=14, axis.title.size = 14)
# axis.size y axis.title.size solo funcionan dentro de "tema_articulo()"
p + tema_articulo2()
p + tema_presentacion()

}
