% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temas_ggplot.R
\name{tema_presentacion}
\alias{tema_presentacion}
\title{Tema con valores predeterminados minimalistas (en mi opinion) adecuados para una presentacion}
\usage{
tema_presentacion(base_size = 24, base_family = "")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
.
}
