% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.back.01.beta.R
\encoding{UTF-8}
\name{trans.back.01.beta}
\alias{trans.back.01.beta}
\title{Re-transformar variable usada en regresión beta, util para proporciones transformadas.}
\usage{
trans.back.01.beta(x)
}
\arguments{
\item{x}{vector que contiene las proporciones transformadas con la función trans.01.beta.}
}
\value{
vector transformado a su escala original.
}
\description{
Permite la re-transformación de las proporciones. Cuando se tienen 0 y 1 absolutos no se puede ajustar un #' modelo de regresion Beta. Arroja una advertencia: invalid dependent variable, all observations must be in
(0, 1). Se recomienda aplicar la transformacion de las proporciones para ajustar modelos con familia beta, propia para proporciones o porcentajes entre 0 y 1 (Douma, 2019).
Esta transformacion permite lidiar con los 0 y 1 absolutos.
}
\examples{
#trans.back.01.beta(x)
#trans.back.01.beta(df$proporciones.trans)
}
\references{
\itemize{
\item Douma, JC, Weedon, JT. Analysing continuous proportions in ecology and evolution:
A practical introduction to beta and Dirichlet regression. Methods Ecol Evol. 2019; 10
:1412– 1430. https://doi.org/10.1111/2041-210X.13234
}
}
