% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.back.arcsine.R
\encoding{UTF-8}
\name{trans.back.arcsine}
\alias{trans.back.arcsine}
\title{Re-transformar variable con arcsine}
\usage{
trans.back.arcsine(x)
}
\arguments{
\item{x}{vector que contiene las proporciones transformadas con arcsine.}
}
\value{
vector transformado a su escala original.
}
\description{
Permite la re-transformación de un vector previamente transformado con Arcsine aplicada a
proporciones. Aplicar la transformación arcsine se ha
sugerido por mucho tiempo como un método estandar al tratar con proporciones/porcentajes (Sokahl & Rohlf, #' 1995). Ver: Sokahl, R., & Rohlf, F. (1995). Biometry (3rd ed.). New York: W. H. Freeman.
}
\examples{
#variable <- seq(1,99, 3)
#var.arcssine <- trans.arcsine(variable)
#trans.back.arcsine(var.arcssine)
}
