% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.back.logit.R
\encoding{UTF-8}
\name{trans.back.logit}
\alias{trans.back.logit}
\title{Re-transformar variable con transformación logit}
\usage{
trans.back.logit(x, a = 0.025)
}
\arguments{
\item{x}{vector que contiene las proporciones transformadas con logit.}

\item{a}{Parémetro fijo de la transformación.}
}
\value{
vector transformado a escala original.
}
\description{
Permite la re-transformación de una variable previamente transformada con logit . Aplicar la transformación logit seguida de una regresión por mínimos cuadrados no lineales sobre  las variables transformadas (Warton & Hui, 2011). Este metodo se ha sugerido por mucho tiempo como un  método para tratar con proporciones/porcentajes. Ver: Warton, D. I., & Hui, F. K. C. (2011). The arcsine is asinine: The analysis of proportions in ecology. Ecology, 92(1), 3–10. https://doi.org/10.1890/10‐0340.1
}
\examples{
#variable <- seq(1,99, 3)
#logit.var<- trans.logit(variable)
#trans.back.logit(logit.var)
}
