% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim_string}
\alias{trim_string}
\title{Trim a scalar string}
\usage{
trim_string(value)
}
\arguments{
\item{value}{Value to trim.}
}
\value{
Character scalar.
}
\description{
Converts a value to a scalar character string and trims surrounding whitespace.
}
