% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{truncate_string}
\alias{truncate_string}
\title{Truncate a string}
\usage{
truncate_string(value, width = 200L)
}
\arguments{
\item{value}{Value to truncate.}

\item{width}{Maximum display width.}
}
\value{
Character scalar.
}
\description{
Shortens a string to a target display width and appends an ellipsis when needed.
}
