% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{ttrssanalytics-package}
\alias{ttrssanalytics}
\alias{ttrssanalytics-package}
\title{ttrssanalytics: TT-RSS News Analytics Pipeline}
\description{
ETL pipeline for extracting and analyzing news from Tiny Tiny RSS.
}
\author{
\strong{Maintainer}: TTRSS Analytics Team \email{ttrss.analytics@example.com}

}
\keyword{internal}
