% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIF_plot.R
\encoding{UTF-8}
\name{VIF_plot}
\alias{VIF_plot}
\title{Grafica el Variance Inflation Factor (VIF) de un modelo}
\usage{
VIF_plot(modelo)
}
\arguments{
\item{modelo}{Modelo, puede ser LM, GLM, Y GLMM.}
}
\value{
VIF plot
}
\description{
Grafica VIF de todas las variables de un modelo.
}
\examples{
#VIF_plot(modelo)
}
