% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_module.R
\name{visualize_ehr_data}
\alias{visualize_ehr_data}
\title{Visualize EHR Data}
\usage{
visualize_ehr_data(ehr_data)
}
\arguments{
\item{ehr_data}{A dataframe containing preprocessed EHR data.}
}
\description{
This function creates visualizations for EHR data, such as histograms, bar charts, etc.
}
