% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amostragem.R
\name{amostra_aleatoria}
\alias{amostra_aleatoria}
\title{Amostragem Aleatória Simples (AAS)}
\usage{
amostra_aleatoria(
  dados,
  n = NULL,
  proporcao = NULL,
  reposicao = FALSE,
  semente = NULL
)
}
\arguments{
\item{dados}{Um data.frame ou vetor representando a população.}

\item{n}{Quantidade exata de observações a sortear.}

\item{proporcao}{Proporção da população a sortear (ex: 0.20 para 20\%).}

\item{reposicao}{Lógico. Se TRUE, permite sortear o mesmo elemento mais de uma vez.}

\item{semente}{Número inteiro para semente aleatória (garante reprodutibilidade).}
}
\value{
Retorna o subconjunto de dados sorteado.
}
\description{
Seleciona uma amostra aleatória simples de um conjunto de dados ou vetor,
com ou sem reposição.
}
\examples{
amostra_aleatoria(iris, n = 30)
amostra_aleatoria(iris, proporcao = 0.2, semente = 123)
}
